requirements.txt
asgiref==3.4.1
Django==3.0.5
django-mongodb-engine==0.6.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
djangotoolbox==1.8.0
djongo==1.3.6
jdatetime==3.6.4
pkg_resources==0.0.0
pymongo==3.12.0
pytz==2021.1
sqlparse==0.2.4
